add default false flag to not delete associated files to delete index command line option to actually delete needs to be added still

+++ zulia-server/src/main/java/io/zulia/server/index/ZuliaIndex.java
@@ -287,13 +287,15 @@ public class ZuliaIndex {
 
 	}
 
-	public void deleteIndex() throws Exception {
+	public void deleteIndex(boolean deleteAssociated) throws Exception {
 
 		unload(true);
 
-		LOG.info("Dropping document storage.");
-		documentStorage.drop();
-		LOG.info("Dropped document storage.");
+		if (deleteAssociated) {
+			LOG.info("Dropping document storage.");
+			documentStorage.drop();
+			LOG.info("Dropped document storage.");
+		}
 
 	}
